refactor: fixture 패키지 이동
