+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.fixture;
 
 import java.math.BigDecimal;
 import java.util.List;
